requirements.txt
python-dotenv~=1.0.1
asyncpg~=0.29.0
sqlmodel~=0.0.19
pydantic~=2.8.2
SQLAlchemy~=2.0.31
black~=24.4.2
isort~=5.13.2
fastapi~=0.111.0
pydantic~=2.8.2
python-json-logger~=2.0.7
starlette~=0.37.2
redis~=5.0.7
uvicorn~=0.30.1
contextvars~=2.4
aiogram~=3.10.0
